// angularUI/src/app/shared/models/entities/user/enums/user-client-permission.enum.ts
export enum UserClientPermissionEnum {
  INSPECTION = 1,
  RECIPE = 2,
  BATCH_DETAIL = 3,
  TEACH = 4,
  USER_MANAGEMENT = 5,
  CAMERA_SETTINGS = 6,
  HARDWARE_SETTINGS = 7,
  PRINTER_SETTINGS = 8,
  SERVER_SETTINGS = 9,
  THIRD_PARTY_SETTINGS = 10,
  SYSTEM_SETTINGS = 11,
  TEST_IO = 12,
  FAULT_IMAGES = 13,
  PRODUCTION_REPORT = 14,
  AUDIT_REPORT = 15,
  HELP = 16,
  INFO = 17,
  RECONCILIATION = 18,
  WAREHOUSE = 19,
  SERVER_ACCESS = 20,
  BATCH_ACCESS = 21,
  BATCH_UPLOAD_ACCESS = 22
}